requirements.txt
discord-py-interactions
python-dotenv
requests
saucenao_api
imagehash
bs4
interactions-files
dinteractions-Paginator
mysql-connector
emoji==1.7.0
genshin
